fix(mtp dp): use int64 init idle batch inputs id in mtp dp mode to fix dtype mismatch in allreduce

+++ python/sglang/srt/speculative/eagle_info_v2.py
@@ -454,7 +454,7 @@ class EagleVerifyInputV2Mixin:
         """
         device = batch.device
         if batch.forward_mode.is_idle():
-            predict = torch.empty(0, dtype=torch.int32, device=device)
+            predict = torch.empty(0, dtype=torch.int64, device=device)
             num_correct_drafts = torch.empty(0, dtype=torch.int32, device=device)
             accept_index = torch.empty(0, dtype=torch.int32, device=device)
             return predict, num_correct_drafts, accept_index
